+++ evaluation.py
@@ -1,39 +1,5 @@
-from util import isNumber
+from util import convertToList
 from validation import validateExpression
-
-
-def convertToList(input_to_calc: str):
-    last_num = False
-    next_unary = True  # if true, next operator will be applied to number (for negation)
-    negate = False
-    expr = []
-    for char in input_to_calc:
-        if isNumber(char):
-            if last_num:
-                if negate:
-                    expr[-1] = expr[-1] * 10 - (ord(char) - 48)  # convert string to number, subtract from end of number in list
-                else:
-                    expr[-1] = expr[-1] * 10 + (ord(char)-48)    # convert string to number, add onto end of number in list
-            else:
-                last_num = True
-                if negate:
-                    expr.append((ord(char)-48)*-1)
-                else:
-                    expr.append(ord(char)-48)     # convert from string to number and add to list
-            next_unary = False  # if operator follows number, it is not unary
-            negate = False
-        elif char == ' ':
-            last_num = False
-        elif char == '-' and next_unary:
-            negate = not negate
-        elif char == ')':
-            last_num = False
-            expr.append(char)
-        else:
-            last_num = False
-            next_unary = True
-            expr.append(char)
-    return expr
 
 
 def getPrecedence(token):
@@ -60,18 +26,18 @@ def applyOp(val1, op, val2):
         return val1 / val2  # shouldn't be any other possible values as validChar function handles those cases
 
 
-def act(val_stack, op_stack):
-    op = op_stack.pop()
+def rep(val_stack, op_stack):
     val2 = val_stack.pop()
     if val2 == 0:
         return "Error: divide by zero"
     val1 = val_stack.pop()
+    op = op_stack.pop()
     res = applyOp(val1, op, val2)
     val_stack.append(res)
     return None
 
 
-def evaluate(expr):
+def evaluateExpression(expr):
     val_stack = []
     op_stack = []
     for token in expr:
@@ -81,19 +47,19 @@ def evaluate(expr):
             op_stack.append(token)
         elif token == ")":
             while len(op_stack) != 0 and op_stack[-1] != "(":
-                err = act(val_stack, op_stack)
+                err = rep(val_stack, op_stack)
                 if err is not None:
                     return err
             op_stack.pop()  # discard "("
         else:
             while len(op_stack) != 0 and op_stack[-1] != '(' and getPrecedence(op_stack[-1]) >= getPrecedence(token):
-                err = act(val_stack, op_stack)
+                err = rep(val_stack, op_stack)
                 if err is not None:
                     return err
             op_stack.append(token)
 
     while len(op_stack) != 0:
-        err = act(val_stack, op_stack)
+        err = rep(val_stack, op_stack)
         if err is not None:
             return err
     return val_stack.pop()
@@ -104,5 +70,5 @@ def calculate(stuff_to_calculate: str):
     list_to_calculate = convertToList(stuff_to_calculate)
     validation_result = validateExpression(list_to_calculate)
     if validation_result is None:
-        return evaluate(list_to_calculate)
+        return evaluateExpression(list_to_calculate)
     return validation_result
