+++ validation.py
@@ -1,27 +1,10 @@
-from util import isNumber, isOperation
-
-
-# def validateExpression(expression: str):
-#     last_unit_is_op = True
-#     i = 0
-#     while i < len(expression):
-#         if isOperation(expression[i]):
-#             if last_unit_is_op and expression[i] != '-':
-#                 return False, i
-#             last_unit_is_op = True
-#         elif isNumber(expression[i]):
-#             if not last_unit_is_op:
-#                 return False, i
-#             last_unit_is_op = False
-#             i = findEndOfNumber(expression, i)
-#         i = i + 1
-#     return not last_unit_is_op, len(expression)
+from util import isOperator
 
 
 def validateExpression(expression: []):
-    if isOperation(expression[0]):
+    if isOperator(expression[0]):
         return "Error: starts with operator"
-    if isOperation(expression[-1]):
+    if isOperator(expression[-1]):
         return "Error: ends with operator"
 
     last_op = ' '
@@ -37,7 +20,7 @@ def validateExpression(expression: []):
             last_op = ''
             last_num = i
             last_bracket = ''
-        elif isOperation(i):
+        elif isOperator(i):
             if last_op != '':
                 return "Error: two operators in a row: " + last_op + " and " + i
             if last_bracket == '(':
@@ -68,22 +51,3 @@ def validateExpression(expression: []):
         return "Error: open right bracket"
 
     return None
-
-
-validChars = {'+', '-', '/', '*', '^', '(', ')', ' '}
-
-
-def isValidChars(input_to_calc: str) -> bool:
-    for tmpChar in input_to_calc:
-        special = validChars.intersection({tmpChar})
-        if len(special) == 0 and not isNumber(tmpChar):
-            return False
-    return True
-
-
-def findEndOfNumber(exp: str, curI: int) -> int:
-    for i in range(curI, len(exp)):
-        if not isNumber(exp[i]):
-            return i-1
-    return len(exp)-1
-
