+++ validation_test.py
@@ -1,4 +1,4 @@
-from validation import validateExpression, findEndOfNumber
+from validation import validateExpression
 from testing_kit import Test
 
 
@@ -28,22 +28,3 @@ class inp:
     def __init__(self,exp,i):
         self.exp = exp
         self.i = i
-
-
-def test_findEndOfNumber():
-    tests = [Test("1 Digit", inp("1", 0), 0),
-             Test("2 Digits", inp("10", 0), 1),
-             Test("2 Digits w/ op", inp("10-", 0), 1),
-             Test("Trips w/Spaces", inp("101 ", 0), 2),
-             Test("No 0 Start", inp("101", 1), 2),
-             Test("Complex", inp(" - 531 124", 3), 5),
-             Test("TwentyOne", inp("21", 0), 1),
-             Test("TwentyOne2", inp(" 21 ", 1), 2)]
-
-    has_failure = False
-    for t in tests:
-        result = findEndOfNumber(t.input.exp,t.input.i)
-        if result != t.output:
-            print("Test", t.name, "failed - Expected:", t.output, "Got:", result)
-            has_failure = True
-    assert(not has_failure)
